final changes for doing decoding and alignment of native recordings
Former-commit-id: 8a3214c07d62b6ec2e0c7637b548cd95e694ed8d [formerly 8a3214c07d62b6ec2e0c7637b548cd95e694ed8d [formerly 8a3214c07d62b6ec2e0c7637b548cd95e694ed8d [formerly b46d9462c04e45ebd84fa13cf0bf97ccfca157f6]]]
Former-commit-id: 04b737be213ec122bc6e25b88817d1356593a61b [formerly 04b737be213ec122bc6e25b88817d1356593a61b [formerly c896bb101031cde485a900cadc4cf4a8c3544784]] [formerly ce2faa9f82b4ec1223b213d3bf39c98bdabacbb9 [formerly ce2faa9f82b4ec1223b213d3bf39c98bdabacbb9 [formerly 6aa12a8b90b1e5f621423252042592fb93bb9702]]]
Former-commit-id: 1314f91c6c2e4642d827618c03eb8a43c8d75862 [formerly ce2faa9f82b4ec1223b213d3bf39c98bdabacbb9]
Former-commit-id: 8910742ccb8ca6b71f3add6d97375564ad2d83cb [formerly 1f2af86c3ae665df4ca7aea460e2b5ab0147da67]
Former-commit-id: 1cf763797ff338bd6d54ced985da8ec4941db688

+++ src/mitll/langtest/server/database/AudioDAO.java
@@ -776,7 +776,7 @@ public class AudioDAO extends DAO {
    * @return AudioAttribute that represents the audio that has been added to the exercise
    * @see mitll.langtest.server.LangTestDatabaseImpl#addToAudioTable(int, String, mitll.langtest.shared.CommonExercise, String, mitll.langtest.shared.AudioAnswer)
    */
-  public AudioAttribute addOrUpdate(int userid, String audioRef, String exerciseID, long timestamp, String audioType, int durationInMillis) {
+  public AudioAttribute addOrUpdate(int userid, String audioRef, String exerciseID, long timestamp, String audioType, long durationInMillis) {
     if (isBadUser(userid)) {
       logger.error("huh? userid is " + userid);
       new Exception().printStackTrace();
@@ -801,7 +801,7 @@ public class AudioDAO extends DAO {
 
       statement.setString(ii++, audioRef);
       statement.setTimestamp(ii++, new Timestamp(timestamp));
-      statement.setInt(ii++, durationInMillis);
+      statement.setInt(ii++, (int)durationInMillis);
 
       statement.setString(ii++, exerciseID);
       statement.setInt(ii++, userid);
@@ -856,7 +856,7 @@ public class AudioDAO extends DAO {
    * @see #addOrUpdate(int, String, String, long, String, int)
    */
   private AudioAttribute getAudioAttribute(int i,
-                                           int userid, String audioRef, String exerciseID, long timestamp, String audioType, int durationInMillis) {
+                                           int userid, String audioRef, String exerciseID, long timestamp, String audioType, long durationInMillis) {
     return new AudioAttribute(i, userid, //id
         exerciseID, // id
         audioRef, // answer
